fixed: routes/api.js, routes/index.js
